backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
torch
torchvision
pillow
opencv-python-headless
pandas
numpy
